+++ Source/DebugUtils/DebugGui.cs
@@ -42,6 +42,24 @@ public static class DebugGui {
     public MethodInfo methodInfo;
   }
 
+  /// <summary>Dumps the objects hierarchy to the logs.</summary>
+  /// <remarks>
+  /// This method recirsively goes down to the all decendants, starting from
+  /// <paramref name="child"/>. The path, however, is shown realtive to <paramref name="root"/>.
+  /// </remarks>
+  /// <param name="root">The root object to make the path from.</param>
+  /// <param name="child">The descendant of <paramref name="root"/> to dump descendants for.</param>
+  public static void DumpHierarchy(Transform  root, Transform child) {
+    if (root != child) {
+      DebugEx.Warning("{0} (localPos: {1}, localRot: {2}, localEuler: {3})",
+                      Hierarchy.MakePath(Hierarchy.GetFullPath(child, parent: root)),
+                      child.localPosition, child.localRotation, child.localRotation.eulerAngles);
+    }
+    for (var i = 0; i < child.childCount; i++) {
+      DumpHierarchy(root, child.GetChild(i));
+    }
+  }
+
   /// <summary>Gets the fields, available for debugging.</summary>
   /// <param name="obj">The instance to get the fields from.</param>
   /// <param name="group">